+++ src/main/java/roomescape/user/controller/ReservationRestController.java
@@ -4,14 +4,12 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.exception.DataNotFoundException;
 import roomescape.user.controller.dto.ReservationRequest;
 import roomescape.user.controller.dto.ReservationResponse;
 import roomescape.user.domain.Reservation;
@@ -23,16 +21,6 @@ import roomescape.user.service.ReservationService;
 public class ReservationRestController {
 
     private final ReservationService reservationService;
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<Void> handleIllegalArgumentException(final IllegalArgumentException e) {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(DataNotFoundException.class)
-    public ResponseEntity<Void> handleDataNotFoundException(final DataNotFoundException e) {
-        return ResponseEntity.notFound().build();
-    }
 
     @PostMapping
     public ResponseEntity<ReservationResponse> persistReservation(
